Cleaning up UI for in-progress swiftui folder
