+++ ios/swiftui/FriendlyChatSwiftUI/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -31,6 +31,7 @@
       "size" : "40x40"
     },
     {
+      "filename" : "firebaseLogo120.png",
       "idiom" : "iphone",
       "scale" : "2x",
       "size" : "60x60"
