+++ ios/swiftui/FriendlyChatSwiftUI/Assets.xcassets/FirebaseGray.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0",
-          "green" : "124",
-          "red" : "245"
+          "blue" : "241",
+          "green" : "239",
+          "red" : "236"
         }
       },
       "idiom" : "universal"
