+++ ios/swiftui/FriendlyChatSwiftUI/Assets.xcassets/FirebaseOrange.colorset/Contents.json
