+++ ios/swiftui/FriendlyChatSwiftUI/Assets.xcassets/FirebaseYellow.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0",
-          "green" : "124",
-          "red" : "245"
+          "blue" : "40",
+          "green" : "202",
+          "red" : "255"
         }
       },
       "idiom" : "universal"
