+++ ios/swiftui/README.md
@@ -2,9 +2,12 @@
 
 This folder is in progress. It will contain an update to the Swift version of this codelab for SwiftUI, better aligning with the Android version of this codelab: [Firebase Android Codelab - Build Friendly Chat](https://firebase.google.com/codelabs/firebase-android).
 
-So far it contains: basic UI elements and styling. Follow-up PRs:
+So far it contains: basic UI elements and styling:
 
-1. Adding Firebase Auth, Firestore
+<img src="https://user-images.githubusercontent.com/15021201/185505923-83f7772c-0117-4028-9378-f741522e5699.png" width="300" />
+
+Future PRs:
+1. Adding Firebase Auth, Firestore, Storage
 1. Re-writing codelab to include latest updates + emulators
 1. Removing the current version of this codelab (currently in the swift folder) and replacing it with this version
